requirements.txt
aiohttp
motor
pyyaml